Added lab mr subcommand
create, checkout, list

`lab mr create` same as `lab merge-request`
`lab mr list` lists MRs by updated_at, takes optional page_number
`lab mr checkout` checkout MR as branch using sourceBranchName
